gateway/requirements.txt
flask
requests
PyJWT